+++ include/megaapi_impl.h
@@ -422,8 +422,8 @@ class MegaTransferPrivate : public MegaTransfer, public Cachable
         virtual const char* getAppData() const;
         virtual void setState(int state);
         virtual int getState() const;
-        virtual void setPriority(float p);
-        virtual float getPriority() const;
+        virtual void setPriority(unsigned long long p);
+        virtual unsigned long long getPriority() const;
 
         virtual bool serialize(string*);
         static MegaTransferPrivate* unserialize(string*);
@@ -432,7 +432,7 @@ class MegaTransferPrivate : public MegaTransfer, public Cachable
 		int type;
 		int tag;
         int state;
-        float priority;
+        uint64_t priority;
 
         struct
         {
